requirements.txt
streamlit==1.28.1
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0